backend/pyproject.toml
[tool.poetry]
name = "neu_sa"
version = "0.1.0"
description = ""
authors = ["Dharun Karthick <dharun.ramaraj@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
fastapi = "^0.115.5"
python-dotenv = "^1.0.1"
snowflake-connector-python = "^3.12.3"
argon2-cffi = "^23.1.0"
python-jose = "^3.3.0"
boto3 = "^1.35.71"
pypdf2 = "^3.0.1"
python-multipart = "^0.0.18"
uvicorn = "^0.32.1"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.poetry.scripts]
backend = "neu_sa.fastapp:main"
